Allows GYP_MSVS_VERSION=2017
I am pretty ignorant to node-gyp, but I couldn't get a project to build because it was using 2017 build tools, despite everyone saying that 2017 tools were supported in 3.6.0. I kept getting an error thrown here. So I modified the source and got it to work. 

If there was a reason this shouldn't have been updated, then ignore this PR, but otherwise, I thought I'd share.